requirements.txt
tensorflow
python-multipart
pillow
matplotlib
numpy
streamlit